Refactor command proxies
