--- osu.Game.Rulesets.Osu/Edit/Commands/SliderCommandProxy.cs
@@ -1,9 +1,0 @@
-// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-namespace osu.Game.Rulesets.Osu.Edit.Commands
-{
-    public static class SliderCommandProxy
-    {
-    }
-}
