+++ osu.Game/Rulesets/Objects/SliderPathExtensions.cs
@@ -7,6 +7,7 @@ using osu.Game.Rulesets.Edit;
 using osu.Game.Rulesets.Objects.Types;
 using osu.Game.Screens.Edit;
 using osu.Game.Screens.Edit.Commands;
+using osu.Game.Screens.Edit.Commands.Proxies;
 using osuTK;
 
 namespace osu.Game.Rulesets.Objects
